General cleanup
- Added interactive mode status display
- Added check for initial application status
- Moved state into our command structure
- Factored out scaleUp and scaleDown methods
- Moved start and stop into those methods

There is a change in behavior, in that, if the target application
already has some instances running, we leave those running and just
add the number of instances of the source app. In other words, if
you start with 10 and 0 instances, you end up with 0 and 10. But if
you start with 10 and 5 instances, you end up with 0 and 15. I think
that is probably correct behavior. And it might even be helpful for
recovery purposes (if I abort after half a scaleover, I can now
simply repeat the command and it will pick up where it left off)